requirements.txt
Flask==2.3.3
tensorflow==2.15.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
